+++ packages/figma/src/codegen/targets/react/component/handlers/field.ts
@@ -24,13 +24,13 @@ export const createFieldHeaderHandler = (ctx: ComponentHandlerDeps) => {
   const indicatorHandler = createFieldIndicatorHandler(ctx);
 
   return defineComponentHandler<FieldHeaderProperties>(
-    metadata.privateComponentFieldHeader.key,
+    metadata.componentFieldHeader.key,
     (node, traverse) => {
       const { componentProperties: props } = node;
 
       const [indicator] = findAllInstances<FieldIndicatorProperties>({
         node,
-        key: metadata.privateComponentFieldHeaderIndicator.key,
+        key: indicatorHandler.key,
       });
 
       // only returns some nice common props for Slider, TextField and more
@@ -94,7 +94,7 @@ export const createFieldFooterHandler = (ctx: ComponentHandlerDeps) => {
   const characterCountHandler = createFieldCharacterCountHandler(ctx);
 
   return defineComponentHandler<FieldFooterProperties>(
-    metadata.privateComponentFieldFooter.key,
+    metadata.componentFieldFooter.key,
     (node, traverse) => {
       const { componentProperties: props } = node;
 
@@ -110,7 +110,7 @@ export const createFieldFooterHandler = (ctx: ComponentHandlerDeps) => {
         .with("Character Count", () => {
           const [characterCount] = findAllInstances<FieldCharacterCountProperties>({
             node,
-            key: metadata.privateComponentFieldFooterCharacterCount.key,
+            key: characterCountHandler.key,
           });
 
           return {