+++ packages/figma/src/entities/data/__generated__/components/index.mjs
@@ -1,3 +1,28 @@
+export const privateTemplateAddressPickerField = {
+  "name": "privateTemplateAddressPickerField",
+  "key": "4af06df28eca43fe2be5fe5ba5e6019587de9fac"
+};
+
+export const privateTemplateDatePickerField = {
+  "name": "privateTemplateDatePickerField",
+  "key": "c161d1326a1087258e4f762aa3c378c098308d98"
+};
+
+export const privateTemplateSelectField = {
+  "name": "privateTemplateSelectField",
+  "key": "a2138764f60a9b5a35e22ff40bc6cd701c660260"
+};
+
+export const privateTemplateTimePickerField = {
+  "name": "privateTemplateTimePickerField",
+  "key": "e38df17cf1e0f96e09774b015739dfde30d46115"
+};
+
+export const privateComponentBottomSheetContentsPlaceholder = {
+  "name": "privateComponentBottomSheetContentsPlaceholder",
+  "key": "e68b006d572300d3c987776192c8ab387fa45e05"
+};
+
 export const privateComponentField = {
   "name": "privateComponentField",
   "key": "96f0d114c3ec7826b22531502f05e82404835df4",
@@ -12,6 +37,21 @@ export const privateComponentField = {
       "type": "BOOLEAN"
     }
   }
+};
+
+export const privateComponentSelectBoxItemCustomContent = {
+  "name": "privateComponentSelectBoxItemCustomContent",
+  "key": "b327f8f0f74f9af3fded3f22fa4a64a08ed8d7cb"
+};
+
+export const privateComponentSliderItemHandleSliderItemValueIndicator = {
+  "name": "privateComponentSliderItemHandleSliderItemValueIndicator",
+  "key": "cfe6acde9d78c5020d2c84a0fdbd54f75277d231"
+};
+
+export const privateComponentSliderItemTrack = {
+  "name": "privateComponentSliderItemTrack",
+  "key": "dd6ea9f90ca6cabfb2f215155b77877adfaafdc9"
 };
 
 export const privateComponentTopNavigationLeftIconButton = {
@@ -34,6 +74,46 @@ export const templateFilterBar = {
   }
 };
 
+export const componentIcon = {
+  "name": "componentIcon",
+  "key": "4a6fdb3425a44a8bc1fe1502d830fa6e82decef7"
+};
+
+export const templateInformationList = {
+  "name": "templateInformationList",
+  "key": "ea1ff4070a5ef4ada5974dc7030e8f9a1b759cdb"
+};
+
+export const componentFigmaOnly = {
+  "name": "componentFigmaOnly",
+  "key": "a790da2ff1fb6f761abf202034920c1504b8f8f1"
+};
+
+export const componentBottomSheetContentsAgreement = {
+  "name": "componentBottomSheetContentsAgreement",
+  "key": "9d530235407646750ce8298b2815093dffcbbd83"
+};
+
+export const componentBottomSheetContentsFilter = {
+  "name": "componentBottomSheetContentsFilter",
+  "key": "9b50b222edac142fcce609da87250d7778e45ec4"
+};
+
+export const componentBottomSheetContentsImage = {
+  "name": "componentBottomSheetContentsImage",
+  "key": "b0b3e0f267ee67e93d6af524ce9d45764027507c"
+};
+
+export const componentBottomSheetContentsRange = {
+  "name": "componentBottomSheetContentsRange",
+  "key": "0e8d8770bacfb2567a054e9c2159fe5dad6a21ba"
+};
+
+export const componentBottomSheetContentsSorting = {
+  "name": "componentBottomSheetContentsSorting",
+  "key": "61f7b8ff0a1a374a10bb1a6a7a31f379863f2ac8"
+};
+
 export const componentChipSuffixIcon = {
   "name": "componentChipSuffixIcon",
   "key": "2f79e3c5a78315c854d7bd4499d142cfcc94548f",
@@ -42,6 +122,16 @@ export const componentChipSuffixIcon = {
       "type": "INSTANCE_SWAP"
     }
   }
+};
+
+export const componentImageFrameBadge = {
+  "name": "componentImageFrameBadge",
+  "key": "6a1feb47139040d6f7522528f7c91bf4fe7bcc84"
+};
+
+export const componentImageFrameCustom = {
+  "name": "componentImageFrameCustom",
+  "key": "60900f1f80431faf16f78e0570848d4656076990"
 };
 
 export const componentImageFrameIcon = {
@@ -64,6 +154,41 @@ export const componentImageFrameOverlayIndicator = {
   }
 };
 
+export const componentListItemDetailCustom = {
+  "name": "componentListItemDetailCustom",
+  "key": "d56900622b320739b5ce9dc9b12af1bcd1f6eeb4"
+};
+
+export const componentListItemDetailSubText_42a = {
+  "name": "componentListItemDetailSubText",
+  "key": "42ae9274fba94cb3b02c4e64b6584e5d03b73835"
+};
+
+export const componentListItemDetailSubText_07e = {
+  "name": "componentListItemDetailSubText",
+  "key": "07ea4a584c3fc5862e842efb5a151aff0059ee76"
+};
+
+export const componentListItemDetailTagGroup = {
+  "name": "componentListItemDetailTagGroup",
+  "key": "8e4743cce1ff6ac9e85269de6ce0ecff0ade06cb"
+};
+
+export const componentListItemPrefixAvatar = {
+  "name": "componentListItemPrefixAvatar",
+  "key": "27e33754113178be97e07195528c4ea020b3d3b7"
+};
+
+export const componentListItemPrefixCheckbox = {
+  "name": "componentListItemPrefixCheckbox",
+  "key": "563275de82ea1282cece0c35c0cd8d1625bc3a9d"
+};
+
+export const componentListItemPrefixCustom = {
+  "name": "componentListItemPrefixCustom",
+  "key": "b8059f5e0f85e0745fc61ff70f04571177c2cdfc"
+};
+
 export const componentListItemPrefixIcon = {
   "name": "componentListItemPrefixIcon",
   "key": "6c03690f1ce9f6c8b2fcdf4a7c57784f6cca12b9",
@@ -72,6 +197,21 @@ export const componentListItemPrefixIcon = {
       "type": "INSTANCE_SWAP"
     }
   }
+};
+
+export const componentListItemPrefixImage = {
+  "name": "componentListItemPrefixImage",
+  "key": "d06216ff143a960844799c0b8f9212628f78c69d"
+};
+
+export const componentListItemPrefixRadiomark = {
+  "name": "componentListItemPrefixRadiomark",
+  "key": "51f7c0917ebc559d81e63d0639cb632a792f40de"
+};
+
+export const componentListItemSuffixCheckbox = {
+  "name": "componentListItemSuffixCheckbox",
+  "key": "385ba8d607029e15e0d38ab415f783016488b185"
 };
 
 export const componentListItemSuffixChevron = {
@@ -97,6 +237,11 @@ export const componentListItemSuffixChevronWithText = {
   }
 };
 
+export const componentListItemSuffixCustom = {
+  "name": "componentListItemSuffixCustom",
+  "key": "26b86c9f8965d38aa5a1181a5cdc89fa487988d1"
+};
+
 export const componentListItemSuffixIcon = {
   "name": "componentListItemSuffixIcon",
   "key": "b7582e74a4bae29df8bc3f81368e528701a75855",
@@ -107,6 +252,16 @@ export const componentListItemSuffixIcon = {
   }
 };
 
+export const componentListItemSuffixRadiomark = {
+  "name": "componentListItemSuffixRadiomark",
+  "key": "09871d64c5c30407da586fb34425c2e83e147c81"
+};
+
+export const componentListItemSuffixSwitch = {
+  "name": "componentListItemSuffixSwitch",
+  "key": "0c26bd64e117e168b06eea69be903e4be762a728"
+};
+
 export const componentPageBannerSuffixAction = {
   "name": "componentPageBannerSuffixAction",
   "key": "1bbd6fff9a32b4211bbe3eeb09fde4e12e87caed",
@@ -115,4 +270,109 @@ export const componentPageBannerSuffixAction = {
       "type": "TEXT"
     }
   }
+};
+
+export const componentPageBannerSuffixCustom = {
+  "name": "componentPageBannerSuffixCustom",
+  "key": "40f2b4754bf16fa268941d03499dea575baa7e26"
+};
+
+export const componentPageBannerSuffixDismiss = {
+  "name": "componentPageBannerSuffixDismiss",
+  "key": "7bdf687e01996f3582b4056954fcb65f0bc67b2f"
+};
+
+export const componentPageBannerSuffixChevron = {
+  "name": "componentPageBannerSuffixChevron",
+  "key": "6f570f3b2b50649b0fd81190ebb42604ae3aa3a5"
+};
+
+export const componentResultSectionAssetTown = {
+  "name": "componentResultSectionAssetTown",
+  "key": "fb790e6aac3dea61df651faeb68078835c3e59ab"
+};
+
+export const componentResultSectionBridgeLottie = {
+  "name": "componentResultSectionBridgeLottie",
+  "key": "340c0c7a24b5a4c215b9e699d7af44578939485d"
+};
+
+export const componentResultSectionCompleteLottie = {
+  "name": "componentResultSectionCompleteLottie",
+  "key": "8c337e54aa38b7eb1eec9281c4934c5f0f87124c"
+};
+
+export const componentResultSectionCustom = {
+  "name": "componentResultSectionCustom",
+  "key": "c99bcc071b3ec4243ecdc1502545f427250a7286"
+};
+
+export const componentResultSectionFailLottie = {
+  "name": "componentResultSectionFailLottie",
+  "key": "ebace2f76203529076734076af40fbd04146b724"
+};
+
+export const componentResultSectionIconEmptyChat = {
+  "name": "componentResultSectionIconEmptyChat",
+  "key": "ed86a05df0ddeea4a84008ace70f750f5d1e46c9"
+};
+
+export const componentResultSectionIconEmptyList = {
+  "name": "componentResultSectionIconEmptyList",
+  "key": "99daf8fa5e0eed82c2dab35963f3a8e2f55968a3"
+};
+
+export const componentResultSectionIconEmptyPeople = {
+  "name": "componentResultSectionIconEmptyPeople",
+  "key": "18623f2ffffe641d3f9876189699839f27d279ac"
+};
+
+export const componentResultSectionIconEmptySearch = {
+  "name": "componentResultSectionIconEmptySearch",
+  "key": "4aa9c04ae49806098872f7c856b1925843ca7f4c"
+};
+
+export const componentSelectBoxSuffixCustom = {
+  "name": "componentSelectBoxSuffixCustom",
+  "key": "2ba92fc0cd161281686bf04767e8f2322c8a17f4"
+};
+
+export const componentSelectBoxItemPrefixAvatar = {
+  "name": "componentSelectBoxItemPrefixAvatar",
+  "key": "3d1ea583297f35880d9ec588326d9b82608565cc"
+};
+
+export const componentSelectBoxItemPrefixBadge = {
+  "name": "componentSelectBoxItemPrefixBadge",
+  "key": "d8c60ae70dee3a35795d83a9fd68282c0a90b554"
+};
+
+export const componentSelectBoxItemPrefixCustom = {
+  "name": "componentSelectBoxItemPrefixCustom",
+  "key": "8129b7b7804c8b22cf67d42df2383c1f5152298e"
+};
+
+export const componentSelectBoxItemPrefixIcon = {
+  "name": "componentSelectBoxItemPrefixIcon",
+  "key": "7184053a74eef18503170d08a1bdb74d37cbb65f",
+  "componentPropertyDefinitions": {
+    "Icon#2475:0": {
+      "type": "INSTANCE_SWAP"
+    }
+  }
+};
+
+export const componentSelectBoxItemPrefixImage = {
+  "name": "componentSelectBoxItemPrefixImage",
+  "key": "d4a3d2da63f6da1552f5d8f947105390248c9018"
+};
+
+export const componentSelectBoxItemSuffixCheck = {
+  "name": "componentSelectBoxItemSuffixCheck",
+  "key": "7b4bdf92e6852f0ecf85c1f6e0350ffaf674a172"
+};
+
+export const componentSelectBoxItemSuffixRadiomark = {
+  "name": "componentSelectBoxItemSuffixRadiomark",
+  "key": "cfb712b12fe2805b5594f8c143147c8aa3de490e"
 };