+++ packages/react/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @seed-design/react
 
+## 1.2.3
+
+### Patch Changes
+
+- Updated dependencies [acae645]
+  - @seed-design/react-snackbar@1.0.1
+
+## 1.3.0
+
+### Patch Changes
+
+- 576c2e6: `AlertDialogRoot`, `MenuSheetRoot` 및 `BottomSheetRoot`의 `onOpenChange` 두 번째 인자로 `details`를 제공합니다. `details.reason`과 `details.event`를 사용할 수 있습니다.
+
+  `DialogAction`을 `DialogPrimitive.CloseButton`으로 교체합니다. `AlertDialogAction` `onClick` 핸들러에서 `event.preventDefault()`를 호출하여 닫기 동작을 방지할 수 있습니다. [(예제)](https://seed-design.io/react/components/alert-dialog#prevent-close)
+
+- Updated dependencies [576c2e6]
+  - @seed-design/react-dialog@1.0.2
+  - @seed-design/react-drawer@1.0.8
+
 ## 1.2.1
 
 ### Patch Changes